pyproject.toml
[tool.poetry]
name = "poetry install"
version = "0.1.0"
description = ""
authors = ["Lana <lan2828@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.23"
pydantic-settings = "^2.1.0"
uvicorn = {extras = ["standart"], version = "^0.25.0"}
sqlmodel = "^0.0.14"
pydantic = "^2.5.3"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
sqlalchemy-utils = "^0.41.1"
msgpack-asgi = "^1.1.0"
pyhumps = "^3.8.0"
backoff = "^2.2.1"
aio-pika = "^9.4.0"
fastapi = "^0.109.0"
unidecode = "^1.3.8"
python-slugify = "^8.0.3"
uni-slugify = "^0.1.4"
pydantic-xml = "^2.8.1"
xmltodict = "^0.13.0"
isort = "^5.13.2"
boto3 = "^1.34.44"
httpx = "^0.27.0"
pytz = "^2024.1"
requests = "^2.31.0"
pillow = "^10.3.0"
python-multipart = "^0.0.9"
orjson = "^3.10.0"
pytest-asyncio = "^0.23.6"
redis = "^5.0.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
isort = "^5.13.2"
faker = "^25.2.0"
pre-commit = "^3.7.1"
mypy = "^1.10.0"

[tool.mypy]
files = [
  "app/*.py",
]
plugins = [
  "pydantic.mypy"
]
ignore_missing_imports = true
follow_imports_for_stubs = false
disallow_untyped_decorators = false
disallow_untyped_calls = false
warn_redundant_casts = true
disallow_any_generics = true
check_untyped_defs = true
implicit_reexport = true

[[tool.mypy.overrides]]
module = ".venv,alembic"
ignore_errors = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.autopep8]
max_line_length = 120
ignore = [
  "E501",
  "W6",
]
in-place = true
recursive = true
aggressive = 3

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 120

[tool.black]
preview = true
enable-unstable-feature = ['string_processing']
line_length = 120
target-version = ['py311']
include = '\.pyi?$'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
